+++ packages/mobile/app.config.ts
@@ -9,7 +9,7 @@ export default ({ config }: ConfigContext): ExpoConfig => ({
   },
   name: 'JotBunker',
   slug: 'jotbunker',
-  version: '1.1.2',
+  version: '1.1.3',
   orientation: 'portrait',
   icon: './assets/icon.png',
   userInterfaceStyle: 'dark',